+++ src/main/java/lotto/model/Ball.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 import java.util.HashMap;
 import java.util.Map;
