+++ src/test/java/lotto/model/BallTest.java
@@ -1,5 +1,6 @@
-package lotto;
+package lotto.model;
 
+import lotto.model.Ball;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
