+++ src/test/java/lotto/model/RankTest.java
@@ -1,5 +1,6 @@
-package lotto;
+package lotto.model;
 
+import lotto.model.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
