+++ src/test/java/lotto/model/TicketTest.java
@@ -1,5 +1,7 @@
-package lotto;
+package lotto.model;
 
+import lotto.model.Ball;
+import lotto.model.Ticket;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
